Add server-side hg_handle recycling.  Prior to this change, servers would call hg_core_listen() to malloc 256 hg_handles, pass them to NA_Msg_recv_unexpected() for future incoming requests, and then keep the handles on the context pending_list.   Each time a new request would come in, a handle from the context pending list would be consumed for it and then freed (e.g. hg_core_destroy()) at the end.   When all 256 pending list handles were consumed, the server would allocate a brand new set of 256 handles (in hg_core_listen()).
This change allows mercury to recycle used hg_handles from a server's
pending list back onto the list.   The server will initially allocate
256 handles at start time.  As server requests come in (e.g. recv unexpected)
handles are pulled off the pending list for processing.  When the
processing is complete, the used handle gets passed to
hg_core_trigger_entry().   That routine has been modified to see
if the handle is from the pending list and can be recycled.  If so,
it re-init's it, puts it back on the pending list, and starts a NA
recv unexpected operation on it.   The server will now only allocate
more handles in hg_core_listen() if all 256 are in use (in that case
it will double the number of pending handles to 512 to hopefully better
handle the load).

At HG_Core_context_destroy() time, we switch off recycling in the
hg_context so that the handles can be properly drained off to shut
down the context.

key changes: add "hg_bool_t recycle_pending_handles" to context to
control recycling.   add "hg_bool_t recyclable_handle" to handle to
mark handles that can be recycled.   added an arg to hg_core_create()
to indicate if the handle being created can be recycled.   new
function hg_core_recycle_to_pending_or_destroy() does the actual recycling
(if it fails, it just frees the handle which is safe to do).
hg_core_trigger_entry() now calls hg_core_recycle_to_pending_or_destroy()
instead of hg_core_destroy() when it can recycle handles.