having a "preallocated" flag in the na_bmi_op_id for preallocated op_id's wasn't good enough in the shutdown case, as preallocated bmi op_id's were getting freed by higher level code before they had a chance to clear out of all the callback in the lower level code.
scrap the "preallocated" flag and replace it with an atomic reference
counter (this is the same way the cci na is structured).   the
preallocation of an op_id (if used) counts as an initial reference.
when a preallocate op_id is used, the reference count goes to two.
this allows the references to be dropped in either order (we'll
only free it when the ref count drops to zero, so shutdown frees
memory in the correct order).