NA OFI: fix a potential bug in na_ofi_addr_ht_lookup
When multiple client processes connect/send_msg to server, a race
condition possibly causes the same source to be inserted to AV multiple
times. The original code removes the fi_addr from AV, but if OFI AV
uses one slot for the multiple AV insert of same source, that will cause
some problems.
This patch refines it by checking the fi_addr and only removes it from AV
when the fi_addrs are different.

Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>