Make wires.c:run_client() return true on success, false on error. In wires.c:main(), print a message and clean up on error.
Extract from wireup_start() a subroutine, wireup_send(), for sending
and re-sending requests.  Hang the initial request off of the wire_t
and change the state machine so that it re-sends requests at regular
intervals until an acknowledgement is received.  When an acknowledgement
is received, free the initial request.  Now the wire_t is 64 bytes long
on LP64 targets.

When a wire is released, free its endpoint (if any), free its request
message (if any), reset its remote ID, message length, and expiration.

When a wire is initialized, set its initial previous/next-to-expire
pointers correctly.

In wiring_free_get(), assert that the wire returned is not on the
expiration queue.  Extract a common subexpression.