First draft of UCX plugin, bugs and all, *minus* the wireup code, which resides for now in a different repo, sorry!
This plugin uses the high-level, UCP API to implement send/recv, put/get,
memory registration, and so on.

One tricky thing about UCP is that its addresses (ucp_address_t) can be
huge---292 bytes is not uncommon on our development server---and support
for connectionless communication is virtually non-existent.

I have had to write a little library that establishes and maintains
mappings between remote peers and integer identifiers that identify a
sender.  That is the wireup library.  It resides outside of Mercury, for
now.