Fix some address reference-counting bugs that caused some memory leaks found by Address Sanitizer.  Do not add a reference to the source address to the callback info unless its type is NA_CB_RECV_UNEXPECTED, because NA will only release the source address for that type.
Increase the reference count on an address before starting wireup on it,
and decrease the reference count on an address when its wire dies.

Add new log outlets to track the na_class_t & na_context_t lifecycles.
Add a "noisy" log outlet for the address lifecycle.

Add helper routines addr_incref() and addr_decref() that increase and
decrease, respectively, the reference count on an na_addr_t, and log the
change.  Log reference counts in other places, too.