Fix a couple of bugs affecting wires and their associated data.
recv_callback: if an unexpected message is received for a disconnected
wire, fail "gracefully" with NA_PROTOCOL_ERROR result instead of trying
to treat wire_data_nil like it is an address!

wire_event_callback: don't zero-out a wire cache's content when it closes,
because it may still be in-use.  Wait until it is reclaimed to
zero it out.